Split `logging.go`
